fix: comment deletion bug

+++ app/src/main/java/com/example/cscb07project/systems/CommentAdapter.java
@@ -27,17 +27,19 @@ public class CommentAdapter extends RecyclerView.Adapter<CommentAdapter.CommentV
     private final OnDeleteClickListener deleteClickListener;
     private boolean deleteModeEnabled;
     private UserRepository userRepo;
+    private boolean  isAdmin;
 
     public interface OnDeleteClickListener {
         void onDeleteClick(Comment comment);
     }
 
-    public CommentAdapter(List<Comment> commentList, OnDeleteClickListener deleteClickListener,  boolean deleteModeEnabled, UserRepository userRepo)
+    public CommentAdapter(List<Comment> commentList, OnDeleteClickListener deleteClickListener,  boolean deleteModeEnabled, UserRepository userRepo,boolean isAdmin)
     {
         this.commentList = commentList;
         this.deleteClickListener = deleteClickListener;
         this.deleteModeEnabled = deleteModeEnabled;
         this.userRepo = userRepo;
+        this.isAdmin = isAdmin;
     }
 
     @NonNull
@@ -50,44 +52,50 @@ public class CommentAdapter extends RecyclerView.Adapter<CommentAdapter.CommentV
     @Override
     public void onBindViewHolder(@NonNull CommentViewHolder holder, int position
     ) {
+
+
         Comment comment = commentList.get(position);
+
+        //get user name through user id
         String userId = comment.getUserId();
         holder.commentText.setText(comment.getComment());
-
         userRepo.getUserById(userId).addOnSuccessListener(user->{
             String userName=user.getUsername();
             holder.commentUser.setText(userName);
-
-
-
         });
+
+
+
         int red = Color.rgb(183, 40, 45);
         int grey = Color.rgb(170, 170, 170);
+        if(isAdmin){
+            if (deleteModeEnabled) {
+                holder.deleteCommentButton.setEnabled(true);
+                holder.deleteCommentButton.setClickable(true);
+                holder.deleteCommentButton.setAlpha(1.0f);
 
-        if (deleteModeEnabled) {
-            holder.deleteCommentButton.setEnabled(true);
-            holder.deleteCommentButton.setClickable(true);
-            holder.deleteCommentButton.setAlpha(1.0f);
+                holder.deleteCommentButton.setImageTintList(
+                        ColorStateList.valueOf(red)
+                );
 
-            holder.deleteCommentButton.setImageTintList(
-                    ColorStateList.valueOf(red)
-            );
+                holder.deleteCommentButton.setOnClickListener(v ->
+                        deleteClickListener.onDeleteClick(comment)
+                );
+            }
+            else {
+                holder.deleteCommentButton.setEnabled(false);
+                holder.deleteCommentButton.setClickable(false);
+                holder.deleteCommentButton.setAlpha(0.5f);
 
-            holder.deleteCommentButton.setOnClickListener(v ->
-                    deleteClickListener.onDeleteClick(comment)
-            );
+                holder.deleteCommentButton.setImageTintList(
+                        ColorStateList.valueOf(grey)
+                );
+
+                holder.deleteCommentButton.setOnClickListener(null);
+            }
         }
-        else {
-            holder.deleteCommentButton.setEnabled(false);
-            holder.deleteCommentButton.setClickable(false);
+        else{
             holder.deleteCommentButton.setVisibility(View.GONE);
-            holder.deleteCommentButton.setAlpha(0.5f);
-
-            holder.deleteCommentButton.setImageTintList(
-                    ColorStateList.valueOf(grey)
-            );
-
-            holder.deleteCommentButton.setOnClickListener(null);
         }
 
 
@@ -99,11 +107,14 @@ public class CommentAdapter extends RecyclerView.Adapter<CommentAdapter.CommentV
         notifyDataSetChanged();
     }
 
+    //this method is neccesary for implementing adapter though it has no use
     @Override
     public int getItemCount() {
         return commentList.size();
     }
 
+
+    //this class extends the recyclerview holder and construct a comment view holder for the
     public static class CommentViewHolder extends RecyclerView.ViewHolder {
 
         private final TextView commentUser;
